feat: rewrite without prime button
